# openapi.yaml
openapi: 3.1.0
info:
  title: TutorTalk API
  description: REST API for the TutorTalk personalized language learning platform
  version: 1.0.0
servers:
  - url: /api
    description: API routes
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Difficulty:
      type: string
      enum: [beginner, intermediate, advanced]
    QuestionInputType:
      type: string
      enum: [text, voice, multiple_choice]
...
    Module:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        topic:
          type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        estimated_duration_mins:
          type: integer
        thumbnail_url:
          type: string
          nullable: true
        is_published:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, title, description, topic, difficulty, estimated_duration_mins, is_published, created_at, updated_at]
    Section:
      type: object
      properties:
        id:
          type: string
          format: uuid
        module_id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        order_index:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, module_id, title, content, order_index, created_at, updated_at]
    QuizQuestion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        question_text:
          type: string
        input_type:
          $ref: '#/components/schemas/QuestionInputType'
        options:
          type: array
          items:
            type: string
        correct_answer:
          type: string
        order_index:
          type: integer
      required: [id, question_text, input_type, order_index]
    Quiz:
      type: object
      properties:
        id:
          type: string
          format: uuid
        section_id:
          type: string
          format: uuid
        title:
          type: string
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuizQuestion'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, section_id, title, questions, created_at, updated_at]
...
    ApiError:
      type: object
      properties:
        success:
          type: boolean
          const: false
        error:
          type: string
        details:
          type: string
      required: [success, error]

    # Input schemas
...
    ModuleUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        topic:
          type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        estimated_duration_mins:
          type: integer
        thumbnail_url:
          type: string
        is_published:
          type: boolean
...
    SectionUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        order_index:
          type: integer
...
    QuizUpdate:
      type: object
      properties:
        title:
          type: string
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuizQuestion'
...
paths:
...
  /modules/{id}:
    get:
      operationId: getModule
      summary: Get a single module
      tags: [Modules]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Module details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    const: true
                  data:
                    $ref: '#/components/schemas/Module'
                required: [success, data]
        '404':
          description: Module not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    patch:
      operationId: updateModule
      summary: Update a module
      tags: [Modules]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleUpdate'
      responses:
        '200':
          description: Module updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    const: true
                  data:
                    $ref: '#/components/schemas/Module'
                required: [success, data]
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Module not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    delete:
      operationId: deleteModule
      summary: Delete a module
      tags: [Modules]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Module deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /sections/{id}:
    get:
      operationId: getSection
      summary: Get a single section
      tags: [Sections]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Section details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    const: true
                  data:
                    $ref: '#/components/schemas/Section'
                required: [success, data]
        '404':
          description: Section not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    patch:
      operationId: updateSection
      summary: Update a section
      tags: [Sections]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SectionUpdate'
      responses:
        '200':
          description: Section updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    const: true
                  data:
                    $ref: '#/components/schemas/Section'
                required: [success, data]
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Section not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Conflict (duplicate order_index)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    delete:
      operationId: deleteSection
      summary: Delete a section
      tags: [Sections]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Section deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /quizzes/{id}:
    get:
      operationId: getQuiz
      summary: Get a single quiz
      tags: [Quizzes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Quiz details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    const: true
                  data:
                    $ref: '#/components/schemas/Quiz'
                required: [success, data]
        '404':
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    patch:
      operationId: updateQuiz
      summary: Update a quiz
      tags: [Quizzes]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizUpdate'
      responses:
        '200':
          description: Quiz updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    const: true
                  data:
                    $ref: '#/components/schemas/Quiz'
                required: [success, data]
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    delete:
      operationId: deleteQuiz
      summary: Delete a quiz
      tags: [Quizzes]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Quiz deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # Profiles
...